rename "django_ynh" -> "django2ynh"
see also https://github.com/YunoHost-Apps/django_ynh/pull/13
