+++ Makefile
@@ -31,7 +31,7 @@ update: install-poetry  ## update the sources and installation and generate "con
 lint: ## Run code formatters and linter
 	poetry run flynt --fail-on-change --line_length=${MAX_LINE_LENGTH} .
 	poetry run isort --check-only .
-	poetry run flake8 django_ynh
+	poetry run flake8 django2ynh
 
 fix-code-style: ## Fix code formatting
 	poetry run flynt --line_length=${MAX_LINE_LENGTH} .
