+++ django2ynh/sso_auth/auth_middleware.py
@@ -12,7 +12,7 @@ try:
 except ImportError:
     from django.core.exceptions import SuspiciousOperation
 
-from django_ynh.sso_auth.user_profile import call_setup_user, update_user_profile
+from django2ynh.sso_auth.user_profile import call_setup_user, update_user_profile
 
 
 logger = logging.getLogger(__name__)
