+++ django2ynh/views.py
