+++ manifest.json
@@ -1,16 +1,16 @@
 {
-	"name": "django_ynh",
-	"id": "django_ynh",
+	"name": "django2ynh",
+	"id": "django2ynh",
 	"packaging_format": 1,
 	"description": {
 		"en": "Glue code to package django projects as yunohost apps."
 	},
-	"version": "0.1.5~ynh1",
-	"url": "https://github.com/jedie/django_ynh",
+	"version": "0.2.0rc0~ynh1",
+	"url": "https://pypi.org/project/django2ynh/",
 	"license": "GPL-3.0",
 	"maintainer": {
 		"name": "Jens Diemer",
-		"email": "django_ynh@jensdiemer.de"
+		"email": "django2ynh@jensdiemer.de"
 	},
 	"previous_maintainers": [],
 	"requirements": {
@@ -26,8 +26,8 @@
 				"name": "domain",
 				"type": "domain",
 				"ask": {
-					"en": "Choose a domain for django_ynh",
-					"fr": "Choisissez un domaine pour django_ynh"
+					"en": "Choose a domain for django2ynh",
+					"fr": "Choisissez un domaine pour django2ynh"
 				},
 				"example": "domain.org"
 			},
@@ -35,18 +35,18 @@
 				"name": "path",
 				"type": "path",
 				"ask": {
-				    "en": "Choose a path for django_ynh",
-				    "fr": "Choisissez un chemin pour django_ynh"
+				    "en": "Choose a path for django2ynh",
+				    "fr": "Choisissez un chemin pour django2ynh"
 				},
-				"example": "/django_ynh",
-				"default": "/django_ynh"
+				"example": "/django2ynh",
+				"default": "/django2ynh"
 			},
 			{
 				"name": "admin",
 				"type": "user",
 				"ask": {
-				    "en": "Choose an admin user for django_ynh",
-				    "fr": "Choisissez l'administrateur pour django_ynh"
+				    "en": "Choose an admin user for django2ynh",
+				    "fr": "Choisissez l'administrateur pour django2ynh"
 				},
 				"example": "johndoe"
 			},
@@ -54,8 +54,8 @@
 				"name": "is_public",
 				"type": "boolean",
 				"ask": {
-					"en": "Should django_ynh be public accessible?",
-					"fr": "django_ynh doit-il être accessible au public ?"
+					"en": "Should django2ynh be public accessible?",
+					"fr": "django2ynh doit-il être accessible au public ?"
 				},
 				"help": {
 					"en": "Any YunoHost user and anonymous people from the web will be able to access the application",
