+++ scripts/restore
@@ -118,7 +118,7 @@ ynh_restore_file --origin_path="/etc/logrotate.d/$app"
 #=================================================
 # GENERIC FINALIZATION
 #=================================================
-# START django_ynh
+# START django2ynh
 #=================================================
 ynh_script_progression --message="Starting a systemd service..." --weight=5
