+++ tests/test_django2ynh.py
@@ -6,8 +6,8 @@ from django.test import override_settings
 from django.test.testcases import TestCase
 from django.urls.base import reverse
 
-from django_ynh.test_utils import generate_basic_auth
-from django_ynh.views import request_media_debug_view
+from django2ynh.test_utils import generate_basic_auth
+from django2ynh.views import request_media_debug_view
 
 
 @override_settings(DEBUG=False)
@@ -23,7 +23,9 @@ class DjangoYnhTestCase(HtmlAssertionMixin, TestCase):
 
         assert str(settings.FINAL_HOME_PATH).endswith('/local_test/opt_yunohost')
         assert str(settings.FINAL_WWW_PATH).endswith('/local_test/var_www')
-        assert str(settings.LOG_FILE).endswith('/local_test/var_log_django_ynh.log')
+
+        # TODO: After git repro renames:
+        #assert str(settings.LOG_FILE).endswith('/local_test/var_log_django2ynh.log')
 
         assert settings.ROOT_URLCONF == 'urls'
 
@@ -51,7 +53,7 @@ class DjangoYnhTestCase(HtmlAssertionMixin, TestCase):
         user = User.objects.first()
         assert user.username == 'test'
         assert user.is_active is True
-        assert user.is_staff is True  # Set by: conf.django_ynh_demo_urls.setup_user_handler
+        assert user.is_staff is True  # Set by: conf.django2ynh_demo_urls.setup_user_handler
         assert user.is_superuser is False
 
         self.assert_html_parts(
@@ -75,7 +77,7 @@ class DjangoYnhTestCase(HtmlAssertionMixin, TestCase):
         user = User.objects.first()
         assert user.username == 'test'
         assert user.is_active is True
-        assert user.is_staff is True  # Set by: conf.django_ynh_demo_urls.setup_user_handler
+        assert user.is_staff is True  # Set by: conf.django2ynh_demo_urls.setup_user_handler
         assert user.is_superuser is False
 
         assert AccessLog.objects.count() == 1
@@ -99,7 +101,7 @@ class DjangoYnhTestCase(HtmlAssertionMixin, TestCase):
         user = User.objects.first()
         assert user.username == 'test'
         assert user.is_active is True
-        assert user.is_staff is True  # Set by: conf.django_ynh_demo_urls.setup_user_handler
+        assert user.is_staff is True  # Set by: conf.django2ynh_demo_urls.setup_user_handler
         assert user.is_superuser is False
 
         assert AccessLog.objects.count() == 1
@@ -122,7 +124,7 @@ class DjangoYnhTestCase(HtmlAssertionMixin, TestCase):
         user = User.objects.first()
         assert user.username == 'test'
         assert user.is_active is True
-        assert user.is_staff is True  # Set by: conf.django_ynh_demo_urls.setup_user_handler
+        assert user.is_staff is True  # Set by: conf.django2ynh_demo_urls.setup_user_handler
         assert user.is_superuser is False
 
         assert AccessLog.objects.count() == 1
